requirements.txt
google-adk>=0.6.0
google-genai>=1.0.0
python-dotenv>=1.0.1
